+++ Presentation/Nop.Web/Administration/Controllers/BlogController.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Web;
 using System.Web.Mvc;
 using Nop.Admin.Models.Blogs;
 using Nop.Core.Domain.Blogs;
@@ -10,15 +9,15 @@ using Nop.Services.Blogs;
 using Nop.Services.Customers;
 using Nop.Services.Helpers;
 using Nop.Services.Localization;
+using Nop.Services.Security;
 using Nop.Web.Framework;
 using Nop.Web.Framework.Controllers;
 using Telerik.Web.Mvc;
-using Nop.Services.Security;
 
 namespace Nop.Admin.Controllers
 {
 	[AdminAuthorize]
-    public class BlogController : BaseNopController
+    public partial class BlogController : BaseNopController
 	{
 		#region Fields
 
@@ -62,15 +61,19 @@ namespace Nop.Admin.Controllers
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageBlog))
                 return AccessDeniedView();
 
-            var blogPosts = _blogService.GetAllBlogPosts(0, null, null, 0, _adminAreaSettings.GridPageSize);
+            var blogPosts = _blogService.GetAllBlogPosts(0, null, null, 0, _adminAreaSettings.GridPageSize, true);
             var gridModel = new GridModel<BlogPostModel>
             {
                 Data = blogPosts.Select(x =>
                 {
                     var m = x.ToModel();
+                    if (x.StartDateUtc.HasValue)
+                        m.StartDate = _dateTimeHelper.ConvertToUserTime(x.StartDateUtc.Value, DateTimeKind.Utc);
+                    if (x.EndDateUtc.HasValue)
+                        m.EndDate = _dateTimeHelper.ConvertToUserTime(x.EndDateUtc.Value, DateTimeKind.Utc);
                     m.CreatedOn = _dateTimeHelper.ConvertToUserTime(x.CreatedOnUtc, DateTimeKind.Utc);
                     m.LanguageName = x.Language.Name;
-                    m.Comments = x.BlogComments.Count;
+                    m.Comments = x.ApprovedCommentCount + x.NotApprovedCommentCount;
                     return m;
                 }),
                 Total = blogPosts.TotalCount
@@ -84,15 +87,19 @@ namespace Nop.Admin.Controllers
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageBlog))
                 return AccessDeniedView();
 
-            var blogPosts = _blogService.GetAllBlogPosts(0, null, null, command.Page - 1, command.PageSize);
+            var blogPosts = _blogService.GetAllBlogPosts(0, null, null, command.Page - 1, command.PageSize, true);
             var gridModel = new GridModel<BlogPostModel>
             {
                 Data = blogPosts.Select(x =>
                 {
                     var m = x.ToModel();
+                    if (x.StartDateUtc.HasValue)
+                        m.StartDate = _dateTimeHelper.ConvertToUserTime(x.StartDateUtc.Value, DateTimeKind.Utc);
+                    if (x.EndDateUtc.HasValue)
+                        m.EndDate = _dateTimeHelper.ConvertToUserTime(x.EndDateUtc.Value, DateTimeKind.Utc);
                     m.CreatedOn = _dateTimeHelper.ConvertToUserTime(x.CreatedOnUtc, DateTimeKind.Utc);
                     m.LanguageName = x.Language.Name;
-                    m.Comments = x.BlogComments.Count;
+                    m.Comments = x.ApprovedCommentCount + x.NotApprovedCommentCount;
                     return m;
                 }),
                 Total = blogPosts.TotalCount
@@ -115,18 +122,17 @@ namespace Nop.Admin.Controllers
             return View(model);
         }
 
-        [HttpPost, FormValueExists("save", "save-continue", "continueEditing")]
+        [HttpPost, ParameterBasedOnFormNameAttribute("save-continue", "continueEditing")]
         public ActionResult Create(BlogPostModel model, bool continueEditing)
         {
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageBlog))
                 return AccessDeniedView();
 
-            //decode body
-            model.Body = HttpUtility.HtmlDecode(model.Body);
-
             if (ModelState.IsValid)
             {
                 var blogPost = model.ToEntity();
+                blogPost.StartDateUtc = model.StartDate;
+                blogPost.EndDateUtc = model.EndDate;
                 blogPost.CreatedOnUtc = DateTime.UtcNow;
                 _blogService.InsertBlogPost(blogPost);
 
@@ -146,14 +152,17 @@ namespace Nop.Admin.Controllers
 
             var blogPost = _blogService.GetBlogPostById(id);
             if (blogPost == null)
-                throw new ArgumentException("No blog post found with the specified id", "id");
+                //No blog post found with the specified id
+                return RedirectToAction("List");
 
             ViewBag.AllLanguages = _languageService.GetAllLanguages(true);
             var model = blogPost.ToModel();
+            model.StartDate = blogPost.StartDateUtc;
+            model.EndDate = blogPost.EndDateUtc;
             return View(model);
 		}
-        
-        [HttpPost, FormValueExists("save", "save-continue", "continueEditing")]
+
+        [HttpPost, ParameterBasedOnFormNameAttribute("save-continue", "continueEditing")]
 		public ActionResult Edit(BlogPostModel model, bool continueEditing)
         {
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageBlog))
@@ -161,14 +170,14 @@ namespace Nop.Admin.Controllers
 
             var blogPost = _blogService.GetBlogPostById(model.Id);
             if (blogPost == null)
-                throw new ArgumentException("No blog post found with the specified id");
-
-            //decode body
-            model.Body = HttpUtility.HtmlDecode(model.Body);
+                //No blog post found with the specified id
+                return RedirectToAction("List");
 
             if (ModelState.IsValid)
             {
                 blogPost = model.ToEntity(blogPost);
+                blogPost.StartDateUtc = model.StartDate;
+                blogPost.EndDateUtc = model.EndDate;
                 _blogService.UpdateBlogPost(blogPost);
 
                 SuccessNotification(_localizationService.GetResource("Admin.ContentManagement.Blog.BlogPosts.Updated"));
@@ -180,15 +189,17 @@ namespace Nop.Admin.Controllers
             return View(model);
 		}
 
-		[HttpPost, ActionName("Delete")]
-		public ActionResult DeleteConfirmed(int id)
+		[HttpPost]
+		public ActionResult Delete(int id)
         {
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageBlog))
                 return AccessDeniedView();
 
             var blogPost = _blogService.GetBlogPostById(id);
             if (blogPost == null)
-                throw new ArgumentException("No blog post found with the specified id", "id");
+                //No blog post found with the specified id
+                return RedirectToAction("List");
+
             _blogService.DeleteBlogPost(blogPost);
 
             SuccessNotification(_localizationService.GetResource("Admin.ContentManagement.Blog.BlogPosts.Deleted"));
@@ -256,8 +267,14 @@ namespace Nop.Admin.Controllers
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageBlog))
                 return AccessDeniedView();
 
-            var comment = _customerContentService.GetCustomerContentById(id);
+            var comment = _customerContentService.GetCustomerContentById(id) as BlogComment;
+            if (comment == null)
+                throw new ArgumentException("No comment found with the specified id");
+
+            var blogPost = comment.BlogPost;
             _customerContentService.DeleteCustomerContent(comment);
+            //update totals
+            _blogService.UpdateCommentTotals(blogPost);
 
             return Comments(filterByBlogPostId, command);
         }
